Revert "feat: add relay manager with relay health stats"
